requirements.txt
python-jose[cryptography]
passlib[bcrypt] 
bcrypt==4.0.1

# --- LangChain Kütüphaneleri ---
langchain-core
langchain-community
langchain-google-genai
langchain-google-firestore
langchain-google-vertexai
langchain-anthropic
langchain-openai

# HuggingFace / Transformers
huggingface-hub>=0.30.0,<1.0
transformers>=4.43.0

# Dosya Okuyucular (Parsers)
pymupdf
python-docx
openpyxl
python-pptx
google-cloud-vision

# Excel Analizi için
pandas>=2.0.0

# Veritabanı Bağlantı Kütüphaneleri
psycopg2-binary
pymongo
mysql-connector-python
python-dotenv
pyodbc==5.1.0

# Güvenlik
slowapi

# Harici Depolama
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
msal
requests

# Web Framework
fastapi
uvicorn[standard]
gunicorn

# Google Cloud
google-cloud-firestore
google-cloud-storage

# Diğer
pydantic
email-validator
python-multipart
rank-bm25>=0.2.2
sentence-transformers>=2.2.0
tabulate>=0.9.0
